requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
pillow==10.3.0
pytz==2024.1
sqlparse==0.5.0
